improve health check
health check now performs some actual checks

1.  we call the axiom-go validate credentials call

this currently does nothing, but we do it any way, as the client
could improve to do something useful in the future

2.  we run an APL query we expect to fail and validate we get
the expected error back

Because we don't know if the user even has any datasets yet, it's
impossible to create an APL statement we know will be valid.  So,
instead we query with an empty APL statement, which we know is
invalid, and we expect HTTP 400 in response.  By doing this we
ensure that network and authentication issues have been resolved.