fix handling of fields with escaped dots
When converting data from the legacy result format into an
intermediate format, field names containing dots were not
properly escaped.  This resulted in the values not being
found, and that caused us to insert a nil where a string
was expected.  This causes the plugin to panic.

Two fixes:

1.  The data-structure used for the intermediate structure is now
map[string]string which means the 0 value of anything not found
will be the empty string (not nil).  This means if we fail to
find the value for any other reason in the future, it will be ""
and not crash.

2.  I have fixed the field name handling to properly handle dot
when it occurs inside a field name by escaping it the same way.